+++ NEWS.md
@@ -1,6 +1,13 @@
-# gwasRtools 0.1.3.9000 (16 Jan 2024)
+# gwasRtools 0.1.4 (17 Jan 2024)
 
+## Updates
 * `get_loci()` and `get_nearest_gene()` detect BOLT-LMM, SAIGE and REGENIE input automatically. Can be disabled with `detect_headers=FALSE`
+
+## Breaking changes 
+* `get_loci()` option `get_ld_indep` is renamed `ld_clump` to be more consistent with the {ieugwasr} package (https://github.com/MRCIEU/ieugwasr)
+* `get_loci()` option `ld_pruning_r2` is renamed `ld_clump_r2` to be more consistent with the {ieugwasr} package (https://github.com/MRCIEU/ieugwasr)
+* `get_loci()` option `exclude_hla` is renamed `hla_locus` because we are not actually excluding it, just treating it as one continuous locus
+
 
 # gwasRtools 0.1.3 (27 Nov 2023)